[improvements] Improve scripts so they  can be used in the ci to test

+++ console/build_ecosystem.ts
@@ -1,5 +1,5 @@
 import { decoder, encoder } from "../deps.ts";
-import { buildDocs, run } from "./scripts.ts";
+import {buildDocs, exists, run} from "./scripts.ts";
 
 console.log("Building ecosystem...");
 
@@ -24,9 +24,11 @@ for (const i in branches) {
 
 // Create the PM2 ecosystem file so that we can run `pm2 start` after this
 // script is done running
-console.log("Creating ecosystem.config.js for PM2...");
-let configs = decoder.decode(await Deno.readFile("ecosystem.config.sample.js"));
-configs = configs.replace(/\/path\/to\/website/g, Deno.cwd());
-await Deno.writeFile("./ecosystem.config.js", encoder.encode(configs));
+if (await exists("./ecosystem.config.js") === false) {
+  console.log("Creating ecosystem.config.js for PM2...");
+  let configs = decoder.decode(await Deno.readFile("ecosystem.config.sample.js"));
+  configs = configs.replace(/\/path\/to\/website/g, Deno.cwd());
+  await Deno.writeFile("./ecosystem.config.js", encoder.encode(configs));
+}
 
 await run(["git", "checkout", "main"]);
